beauty-booking-backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
psycopg[binary]>=3.1.0
SQLAlchemy>=2.0.35

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# AI Integration
anthropic==0.25.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Validation
email-validator==2.1.0
